+++ alma_client/endpoints/exports.py
@@ -15,10 +15,10 @@ class Exports(Base):
         payout_id: str = None,
         start: datetime = None,
         end: datetime = None,
-        include_child_accounts = False,
-        custom_fields = None,
+        include_child_accounts=False,
+        custom_fields=None,
     ):
-        """ Create a new export"""
+        """Create a new export"""
         data = {
             "type": export_type.value,
             "include_child_accounts": include_child_accounts,
@@ -31,7 +31,7 @@ class Exports(Base):
 
         if end:
             data["end"] = int(end.timestamp())
-        
+
         if custom_fields:
             if isinstance(custom_fields, str):
                 data["custom_fields"] = custom_fields
@@ -39,7 +39,10 @@ class Exports(Base):
                 try:
                     data["custom_fields"] = ",".join(custom_fields)
                 except TypeError as e:
-                    raise TypeError(f"Expected comma-separated string or an iterable yielding strings for custom_fields, got {type(custom_fields)}") from e
+                    raise TypeError(
+                        f"Expected comma-separated string or an iterable yielding "
+                        f"strings for custom_fields, got {type(custom_fields)}"
+                    ) from e
 
         return self.request(self.EXPORTS_PATH).set_body(data).post().expectJson(Export)